move storage related functions to storage.js; move Profile constructor to a independent file
